V3.4.0 usec++14 (#3898)
This reverts commit fff4cc8fc94fb9b1a8424a2b89e9966184c6c4ce.

+++ native/cocos/platform/Image.cpp
@@ -252,7 +252,7 @@ void pngReadCallback(png_structp pngPtr, png_bytep data, png_size_t length) {
 
     if (static_cast<int>(isource->offset + length) <= isource->size) {
         memcpy(data, isource->data + isource->offset, length);
-        isource->offset += length;
+        isource->offset += static_cast<int>(length);
     } else {
         png_error(pngPtr, "pngReaderCallback failed");
     }
@@ -744,9 +744,9 @@ bool Image::initWithPVRv2Data(const unsigned char *data, ssize_t dataLen) {
     _renderFormat = it->second;
 
     //Get size of mipmap
-    _width = CC_SWAP_INT32_LITTLE_TO_HOST(header->width);
-    _height = CC_SWAP_INT32_LITTLE_TO_HOST(header->height);
-    _isCompressed    = true;
+    _width        = CC_SWAP_INT32_LITTLE_TO_HOST(header->width);
+    _height       = CC_SWAP_INT32_LITTLE_TO_HOST(header->height);
+    _isCompressed = true;
 
     //Move by size of header
     _dataLen = dataLen - sizeof(PVRv2TexHeader);
